requirements.txt
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.4
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
